// frontend/src/app/shared/models/usuario.ts
export interface Usuario {
  id: number;
  nome: string;
  registroNis: string;
  cpf: string;
  rg: string;
  rendaFamiliar: number;
  profissao: string;
  ocupacao: string;
  dtNascimento: string;
  atendidoProgramaSocial: boolean;
  programaSocial: string;
  observacao: string;
  endereco: string;
  telefone: string;
  totalPessoas: string;
  necessidadesEspeciais: string;
  nomesIdades: string;
}